fix: Parse `remote_connect_timeout_secs` as a double

+++ libmamba/src/core/curl.cpp
@@ -23,7 +23,7 @@ namespace mamba
             CURL* handle,
             const std::string& url,
             const bool set_low_speed_opt,
-            const long connect_timeout_secs,
+            const double connect_timeout_secs,
             const bool set_ssl_no_revoke,
             const std::optional<std::string>& proxy,
             const std::string& ssl_verify
@@ -123,7 +123,7 @@ namespace mamba
         bool check_resource_exists(
             const std::string& url,
             const bool set_low_speed_opt,
-            const long connect_timeout_secs,
+            const double connect_timeout_secs,
             const bool set_ssl_no_revoke,
             const std::optional<std::string>& proxy,
             const std::string& ssl_verify
@@ -392,7 +392,7 @@ namespace mamba
     void CURLHandle::configure_handle(
         const std::string& url,
         const bool set_low_speed_opt,
-        const long connect_timeout_secs,
+        const double connect_timeout_secs,
         const bool set_ssl_no_revoke,
         const std::optional<std::string>& proxy,
         const std::string& ssl_verify
